+++ lib/vault/api/help.rb
@@ -3,13 +3,23 @@ require_relative "../response"
 
 module Vault
   # Help is the response from a help query.
-  class Help < Response.new(:help, :see_also); end
+  class Help < Response
+    # @!attribute [r] help
+    #   The help information.
+    #   @return [String]
+    field :help
+
+    # @!attribute [r] see_also
+    #   Additional help documentation to see.
+    #   @return [String]
+    field :see_also
+  end
 
   class Client
     # Gets help for the given path.
     #
     # @example
-    #   Vault.help #=> #<Vault::Help help="..." see_also="...">
+    #   Vault.help("secret") #=> #<Vault::Help help="..." see_also="...">
     #
     # @param [String] path
     #   the path to get help for