+++ lib/vault/api/sys/init.rb
@@ -1,12 +1,23 @@
 require "json"
 
-require_relative "../sys"
-
 module Vault
-  class InitResponse < Response.new(:keys, :root_token); end
+  class InitResponse < Response
+    # @!attribute [r] keys
+    #   List of unseal keys.
+    #   @return [Array<String>]
+    field :keys
 
-  class InitStatus < Response.new(:initialized)
-    alias_method :initialized?, :initialized
+    # @!attribute [r] root_token
+    #   Initial root token.
+    #   @return [String]
+    field :root_token
+  end
+
+  class InitStatus < Response
+    # @!method initialized?
+    #   Returns whether the Vault server is initialized.
+    #   @return [Boolean]
+    field :initialized, as: :initialized?
   end
 
   class Sys
@@ -33,12 +44,15 @@ module Vault
     #   the number of shares
     # @option options [Fixnum] :threshold
     #   the number of keys needed to unlock
+    # @option options [Array] :pgp_keys
+    #   an optional Array of base64-encoded PGP public keys to encrypt shares with
     #
     # @return [InitResponse]
     def init(options = {})
       json = client.put("/v1/sys/init", JSON.fast_generate(
         secret_shares:    options.fetch(:shares, 5),
         secret_threshold: options.fetch(:threshold, 3),
+        pgp_keys:         options.fetch(:pgp_keys, nil)
       ))
       return InitResponse.decode(json)
     end