+++ spec/integration/api/auth_token_spec.rb
@@ -11,6 +11,69 @@ module Vault
         expect(result.auth).to be_a(Vault::SecretAuth)
         expect(result.auth.client_token).to be
       end
+
+      it "creates a new token as a wrapped response" do
+        ttl = 50
+        result = subject.auth_token.create({:wrap_ttl => ttl})
+        expect(result).to be_a(Vault::Secret)
+        expect(result.wrap_info).to be_a(Vault::WrapInfo)
+        expect(result.wrap_info.ttl).to eq(ttl)
+        expect(result.wrap_info.token).to be
+      end
+    end
+
+    describe "#create_orphan" do
+      it "creates an orphaned token" do
+        result = subject.auth_token.create_orphan
+        expect(result).to be_a(Vault::Secret)
+        expect(result.auth).to be_a(Vault::SecretAuth)
+        expect(result.auth.client_token).to be
+      end
+
+      it "creates an orphaned token as a wrapped response" do
+        ttl = 50
+        result = subject.auth_token.create_orphan({:wrap_ttl => ttl})
+        expect(result).to be_a(Vault::Secret)
+        expect(result.wrap_info).to be_a(Vault::WrapInfo)
+        expect(result.wrap_info.ttl).to eq(ttl)
+        expect(result.wrap_info.token).to be
+      end
+    end
+
+    describe "#create_with_role" do
+      it "creates a token attached to a role" do
+        vault_test_client.logical.write("auth/token/roles/default")
+        result = subject.auth_token.create_with_role("default")
+        expect(result).to be_a(Vault::Secret)
+        expect(result.auth).to be_a(Vault::SecretAuth)
+        expect(result.auth.client_token).to be
+      end
+
+      it "creates a new token attached to a role as a wrapped response" do
+        ttl = 50
+        vault_test_client.logical.write("auth/token/roles/default")
+        result = subject.auth_token.create_with_role("default", {:wrap_ttl => ttl})
+        expect(result).to be_a(Vault::Secret)
+        expect(result.wrap_info).to be_a(Vault::WrapInfo)
+        expect(result.wrap_info.ttl).to eq(ttl)
+        expect(result.wrap_info.token).to be
+      end
+    end
+
+    describe "#lookup" do
+      it "retrieves the given token" do
+        result = subject.auth_token.lookup(subject.token)
+        expect(result).to be_a(Vault::Secret)
+        expect(result.data[:id]).to eq(subject.token)
+      end
+    end
+
+    describe "#lookup_self" do
+      it "retrieves the current token" do
+        result = subject.auth_token.lookup_self
+        expect(result).to be_a(Vault::Secret)
+        expect(result.data[:id]).to eq(subject.token)
+      end
     end
 
     describe "#renew_self" do